requirements.txt
python-dotenv
slack-bolt
zammadoo

